Add Parcel configuration and update dependencies
This commit adds a new `.parcelrc` file with configuration for SVG transformers, optimizers, and compressors. It also updates the `package.json` file with new versions of `@parcel/compressor-brotli` and `@parcel/compressor-gzip`. Additionally, it includes changes to the `detail-yuji.html` and `index.html` files, updating the `loading` attribute for images and iframes.